reorganize components, add ConfigNode type
