+++ gridworks/src/components/ConfigViewer/YamlContext.tsx
@@ -1,0 +1,9 @@
+"use client";
+import { createContext } from "react";
+
+export const YamlContext = createContext<{
+  isSelected?: (key: string, value: string) => boolean;
+  onSelectLine?: (key: string, value: string) => void;
+  unsetFields: Set<string>;
+  kind?: string;
+}>({ unsetFields: new Set() });
